Add files via upload
Angioplasty is a minimally invasive endovascular surgical procedure widely used to treat arterial 
atherosclerosis. It consists in the expansion of a polymeric ballon placed into the stenosis tract thanks 
to a catheter guide, in order to compress the sclerotic plaque and restore the physiological blood flow. 
To prevent re-stenosis and support the damaged vessel walls, the procedure can be performed in 
combination with stenting, which is the deployment, by plastic deformation, of a metallic reticular tube 
in the damaged tract by means of the same ballon expansion. Materials, shapes and dimension of the 
stent can vary based on treatments and localization of the stenosis. 
The purpose of our study was to practice the use of ABAQUS CAE finite elements software by examining 
the expansion and a subsequent deflation of the balloon inside a stent ring for which both geometries 
and mechanical properties were provided. The parameters of interest are all related to the uncrimping 
process of the stent strut, such as longitudinal shortening, radial recoil and plastic strain field.